requirements.txt
aiohttp==3.9.1
audio_separator==0.19.1
audioread==3.0.1
deepmultilingualpunctuation==1.0.1
Flask==3.0.3
germansentiment==1.1.0
joblib==1.3.2
librosa==0.10.1
Logbook==1.7.0.post0
matplotlib==3.8.2
moviepy==1.0.3
mysql_connector_repackaged==0.3.1
nltk==3.8.1
numpy==2.1.1
pandas==2.2.2
Pillow==10.4.0
plotly==5.18.0
praatio==4.4.0
pydub==0.25.1
pyloudnorm==0.1.1
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.3.2
scipy==1.14.1
selenium==4.24.0
simple_youtube_api==0.2.8
six==1.16.0
sox==1.4.1
tensorboardX==2.6.2.2
tgt==1.5
tiktok_uploader==1.0.16
torch==2.2.0
tqdm==4.66.2
tweepy==4.14.0
vosk==0.3.45
websockets==10.4
yt_dlp==2023.12.30
